refactor: move general reflectometry types out of amor.types
