+++ src/essreflectometry/types.py
@@ -24,9 +24,6 @@ class QData(sciline.Scope[Run, sc.DataArray], sc.DataArray):
     """Theta data transformed to momentum transfer"""
 
 
-QResolution = NewType('QResolution', sc.Variable)
-
-
 class FootprintCorrected(sciline.Scope[Run, sc.DataArray], sc.DataArray):
     """Experiment data corrected by footprint on sample"""
 
@@ -47,6 +44,7 @@ class Normalized(sciline.Scope[Run, sc.DataArray], sc.DataArray):
 
 
 NormalizedIofQ = NewType('NormalizedIofQ', sc.DataArray)
+QResolution = NewType('QResolution', sc.Variable)
 
 
 ''' Parameters for the workflow '''
@@ -56,3 +54,32 @@ QBins = NewType('QBins', sc.Variable)
 
 class Filename(sciline.Scope[Run, str], str):
     """Filename of the raw data"""
+
+
+# TODO What do they mean?
+# Supermirror parameters
+MValue = NewType('MValue', sc.Variable)
+CriticalEdge = NewType('CriticalEdge', sc.Variable)
+Alpha = NewType('Alpha', sc.Variable)
+SupermirrorParameter = TypeVar('SupermirrorParameter', MValue, CriticalEdge, Alpha)
+
+
+class SampleRotation(sciline.Scope[Run, sc.Variable], sc.Variable):
+    """The rotation of the sample / the reference sample."""
+
+
+class BeamSize(sciline.Scope[Run, sc.Variable], sc.Variable):
+    """FWHM of the neutron beam."""
+
+
+class DetectorSpatialResolution(sciline.Scope[Run, sc.Variable], sc.Variable):
+    # TODO
+    """Don't know what this is."""
+
+
+class SampleSize(sciline.Scope[Run, sc.Variable], sc.Variable):
+    # TODO is this radius or total length?
+    """Size of the sample."""
+
+
+Gravity = NewType('Gravity', sc.Variable)
